Added in methods and fixed up UserGroup type
Added in methods to get users based on mod type from a groupID, and the
not moderators. Added in methods to get groupID where a user is a mod,
and groups they aren't. Also maybe fixed up the UserGroup type right, it
was empty with a big TODO in it.